requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain-qdrant>=0.0.1
qdrant-client>=1.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.0.0
sentence-transformers>=2.0.0
requests>=2.28.1
gpt4all==0.1.7
langchain-huggingface